Move funcs that source Markdown bytes to `source.go`
This is purely a move. Functions are byte for byte equivalent at the source level.
